added CustomProtein to __init__ file for model_loaders ; removed redundanttests
