+++ model_loaders/__init__.py
@@ -33,6 +33,7 @@ from .proteins import ProteinMultiTemperature
 from .proteins import ProteinNonLinear
 from .proteins import ProteinAwsem
 from .proteins import ProteinNonBonded
+from.custom_protein import CustomProtein
 
 #load all the helper functions
 from .helper_functions import load_protein
